Add my namespaces page
